Added dynamically sized rendering

+++ include/Line.hpp
@@ -8,19 +8,19 @@ using namespace rgb_matrix;
 namespace matrix_minds {
   class Line {
     private:
-      const float x;
-      const float y;
-      const float height;
-      const float width;
+      const double x;
+      const double y;
+      const double height;
+      const double width;
       const Color color;
 
     public:
-      Line(float x, float y, float height, float width, Color color) : x(x), y(y), height(height), width(width), color(color) {}
+      Line(double x, double y, double height, double width, Color color) : x(x), y(y), height(height), width(width), color(color) {}
       
-      inline float getPosX() const { return this->x; }
-      inline float getPosY() const { return this->y; }
-      inline float getHeight() const { return this->height; }
-      inline float getWidth() const { return this->width; }
+      inline double getPosX() const { return this->x; }
+      inline double getPosY() const { return this->y; }
+      inline double getHeight() const { return this->height; }
+      inline double getWidth() const { return this->width; }
       inline Color getColor() const { return this->color; }
   };
